scripts/requirements.txt
garminconnect>=0.2.8
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
